Split RemoteFailoverTestCase into separate tests. Disable reuseForks, absent another mechanism to isolate ejb client context between tests.
